[Gradle] Cleanup of the test code. Create base class for examples-based tests.
