+++ plugins/gradle/src/test/resources/examples/kotless/shortener/shortener.tf
