+++ plugins/gradle/src/test/resources/examples/kotless/site/site.tf
