+++ plugins/gradle/src/test/resources/examples/ktor/site/site.tf
