+++ plugins/gradle/src/test/resources/examples/spring/shortener/shortener.tf
