+++ plugins/gradle/src/test/resources/examples/spring/site/site.tf
